Bundle exec demo so money gem is loaded
so as to fix the GHActions build which cannot find the Money gem when
running the demo script without the bunlde environment. There is a way
to load the Bundler environment from bin/<script>.rb but am not looking
right now, the other option would be to use something like zeitwork?